Major fix to the seismic display widget.
Previous version had a whole string of problems that have been fixed
in this revision.  Most were linked to fundamental errors in the original
version in the way the widget handled it's Initialize method.  This
version does the right thing and calls Initialize once and only once
when the widget is first created.  When the data X resource changes
(new data is to be displayed) a routine called ReInitialize is called
to make the display consistent with the new data.  This removed annoying
scaling artifacts in the previous version.  Some other changes are:
1.  Added to code to handle the single trace display in a special way.
The previous code had some off by one errors that were especially
nasty in this special case.
2.  Fixed the parameter definitions to allow cleaner initialization of
graphics parameters.  The previous version had some items frozen.
It also contained a lot of debris that was not being used.  I haven't
eliminated all of these because there are some residuals that are
features I may eventually implement.  Forthcoming documentation should
clarify some of this.
3.  Made it easy to change the color of the display markers.  These
now can be changed as X resources.
4.  Fixed some major memory leaks.  There are known remainders but
this is better than previous version.

Note this code retains some scaffolding that has been commented out.
I'm not yet that confident I've found the last bug in this rather
complex piece of software.