Fix g++ complaints about deleting opaque pointers.
Compiling this with g++ revealed some dangerous assumptions in the
destroy procedure for this widget.  All but one were removed
by explicit casts.  A residual error persists for a generic object
whose purpose I don't quite get.  It seems a place for extra stuff
to be added.  Doesn't crash on exit, so presume the warning is
indeed harmless.