Improve error handlers.
ArrivalUpdater.cc
Previous version had an irregular error handling procedure.  Changed so
some problems in updating the database that can corrupt the db will now
cause an exception to be thrown with some hints to the user on how to fix them.
dbgetv errors are flagged but do not generate an exception.  Perhaps the
wrong model, but it is what this version does.

ArrivalUpdater.h
Add documentation for doxygen web pages on seismo.geology.indiana.edu.