Correct less than ideal convergence criteria or weighted stack algorithm.
Previous version had an excessively pessimistic convergence criteria
for the stack.  This and a related error in the convergence calculation
went unnoticed as a kind of cancelling set of errors.  Fortunately,
nothing done previously is invalidated because the loop was always
forced to compute 3 cycles and it seems to converge to high accuracy
in that many cycles anyway.  i.e. the previous version worked but
kind of only accidentally.  Discovered in writing the paper on this
program discussing how this algorithm worked.