Improve and regularize test for inconsistent sample rates.
Removed a dumb floating point equal test n correlation.cc that was causing
dbxcor to abort mysteriously.  In debugging found a similar construct in
array_get_data used to decide if resampling was required.  Decided to regularize
this for the whole library using a generic algorithm (template) placed
in the seispp.h file.