Add support for generic fixed length header data.
This is a new capability in this library under development for some time
but now ready for initial release.  Provides a generic support for any
data format with a fixed length header.  Type examples are SEGY and SAC.
The new code provides a consistent interface into any such data with the
header definition driven by the parameter file HeaderMap.pf.  The object
has a relatively clean interface to ease the process of getting and
putting data to a trace header.  The HeaderMap.pf defines the attribute
namespace for the header.  The ones I've checked in here are or SAC
and SEGY derived from the SAC documentation and PASSCAL's segy definition
respectively.   Well, not quite passcal.  I dropped their extensions, but
this could be easily changed by anyone who wanted to customize this or
redefne a Passcal_SEGY trace data type.