Add automatic memory management code.
The previous version of this object, which provides a simplified
interface to Datascope, had a serious problem with memory leaks.
This happened because the first version was somewhat experimental
but then acquired a life of it's own.  This version implements an
automatic memory management scheme whose concept comes from a commonly
used generic object called a shared pointer.  A shared pointer isn't
used here, but the memory management is implemented the same way.
That is, each copy of the DatascopeHandle contain a pointer to an
STL multiset that keeps track of a list of open tables.  When
a handle goes out of scope and is destroyed the destructor erases
it's entry from this multiset.  This allows a simple algorithm to
decide when the memory for a view should be automatically released.
That is, when the last entry for a table is removed from the multiset
(count goes to 0) dbfree is called on that table.

This version has implmented this, but be warned it hasn't been fully
checked to verify if it truly eliminates all memory leaks.  It passed
my test program that exercises all it's elements, but I have not run
it through a memory checker because one wasn't available on the machine
I used to add this feature.  More changes are possible, although
the test program suggest strongly this is working as designed.
One change is certain.  In this version I intentionally left in the
debug scaffolding but the debug code is commented ount.