Fix additional memory leaks and add a test to prevent a subtle error.
This plugs additional leaks found in dsseis.c function linked to
calls to dbjoin with unassociated dbfree calls.  I think I've plugged
all the leaks in this version, but totalview thinks I still have a
small problem.  Run in progess suggest if so it isn't very serious.

In tracking this problem I encountered a nasty potential error in
the little gizmo that deals with bundling three-component data.
It was failing mysteriously on some data.  The problem was tracked
to duplicate tags of BHZ in the default map which caused the
mapper to get confused because the second entry replaced the
first in the map container.  Fixed by having the constructor
throw and exception if a channel code is ever repeated in the
definition.  This seems to pretty effectively prevent the problem
from surfacing mysteriously as it did for me.