Use deployment table to create VND for DMC.
Reverse conversion, between Excel VND files and deployment table,
may be possible with vnd2deployment, but not all information will
be included, and the input Excel/VND format may be fluid.

See also soon to be checked in mk_dmc_files for an "automated"
way to generate VND or dataless, track the generated files,
and ship to DMC using orbxfer2.