Generate and track VNDs or dataless SEED files.
Allows user to comment on why files were generated
or changed and what files were changed.  The files
generated, orb that was used for orbxfer2, and time
file was sent are tracked in the dmcfiles extension
table.

Note that the DMC requested IMS short format phase
pick files are not tracked in this table, but in the
dmcbull extension table (see db2ims and ims2dmc).

Beware of attempting to send files to an orb with
orbxfer where no one is connected.  May need to modify
your orbxfer2.pf file.