Demote and exception to allow a rare error to be handled cleanly.  Document same.
I've been annoyed this little feature for months and finally figured out
what happened.  When constructors for seismograms read data with an internal
gap the results are correctly marked as a gap.  However, I had forgotten
a feature I had included myself and not documented well.  That is, what does
one do with a gap that is at the beginning or end of a desired time window?
Long ago I apparently concluded the right approach is to note mark this
as a gap but require instead that all handlers must alway look carefully
at start and endtiem for consistency.  This was intentional, I now remember,
because if you are working with segmented data this isn't necessarily correct
to mark this situation as a gap.  Failure to remember this design feature
caused a rare error in dbxcor linked to the cross-correlation function
which is being checked in here.  That is, the correlation function used
to throw an exception when the length of the seismogram to be correlated
was less than the correlator length.  (a cryptic complaint about lx<ly
which wasn't good either)  Now this condition is handled the same way
gaps are handled.  That is, the correlator returns a cross-correlation
function that has it's header marked dead.  MultichannelCorrelator already
handles this correctly as this is how it dealt with data having marked
gaps before.

I updated the web documentation to clarify this as well.  Also note
this now tests the SEISPP_verbose variable and only posts a warning
if verbose is on.  Otherwise what used to be an exception will now
be handled silently.