A very very extensive set of changes to this library.  Anyone using this library explicitly or implicitly needs to be warned that I can almost guarantee any application using it will change. By far the most important is that the seispp_attribute_map.pf which will be installed on building this will cause applications using the library to break unless they are recompiled and linked with this new code.  I hated to do this, but it had to be done sometime.
The primary motivation behind this rather large set of changes
was a need for a more general way to handle mapping of external
database attributes to header variable names.  The previous version
used a frozen namespace linked to the above mentioned pf that was
to inflexible to work correctly when I encountered the need to
mix schema  definitions in the same program.  To deal with this
I created an alias Tbl in the parameter file that lists what
tables are to be considered to match a simple attribute name.
For example, samprate in the new pf file is allowed to be an
alias for wfdisc.samprate or wfprocess.samprate.  The alias
can also be something different.  For example, the parameter file
now defines atime to be an alias for arrival.time.

The worst news about these changes is they are not fully and
completely tested.  I'm confident I squashed the worst bugs, but
I have no reason to believe this is bug free.  I'm inflicting
this on my colleagues because I am forced to check these changes
in to allow me to do some further testing on MacOS and SunOS.
I would especially advise anyone using dbxcor to expect some
transients if you rebuild contrib with this library.  My apologies
for any trouble this causes in the short term.