Initial check in of mchedr conversion script.
This script takes a mchedr (USGS) format file and extracts
origin and arrival information and adds it to a css3.0 database.

It has not been thoroughly tested, but appears to work on USGS
mine and explosion data.  More work is needed on converting to
elog routines, reducing verbosity, and possibly creating moment
and centryd tables.

Note that I may abandon this script as I plan on future development
of more generalized bulletin grab-and-convert scripts.