Make the database constructor for this object more robust.
The changes here are a cascade from the new alias mechanism
I added for extracting attributes from a database.  The previous
version of the db constructor dogmatically insisted on extracting
three required attributes (ns,t0, and samprate) by fixed names
from the database.  This collided with the new alias mechanism
which to date is implemented in the Metadata object.  This
fix checks to see if these attributes were loaded as Metadata
and uses the older method only when that process fails.