Fix a very subtle memory error.
The previous version would cause a seg fault that was platform dependent.
i.e. worked on linux and solaris, but failed under MacOS.  Problem was
traced to the aliasmap variable going out of scope and being deleted
leading to a hanging pointer.  The differences in behaviour must be
related to how different compilers time such deletions as the offending
portions were probably nanoseconds apart in execution.