Add manual correction and polarity changing to GUI.
This implements the last major elements of the GUI for this program
that I have planned.  This includes two major additions:
1.  A polarity changing mode enabled by a button on the main
display.  Once enabled once can click and switch polarity on
any trace in the display.
2.  enable and disable a manual picking mode.  This is functionally
similar in some ways to magnify in dbpick, but works quite
differently.  This launches a pop up shell that can be used
to repair cycle skips.  This functionality is very important
in working with narrow band data like teleseismic P seen only
at short periods.

the man page needs and update for these two new features, but until
I get more feedback on how this works I'm avoiding writing it.  A
key reason is there are some confusing sign issues in this display
that could probably go either way.