Fix a very nasty, but subtle bug in interpolator.
Finding this one was one of the most difficult debugging
tasks I have ever faced.  Core routine called billions of
times in my migration program that would fail somewhere
around 1% or less of the time.  Problem was traced to
a very subtle flaw in the algorithm.  The original version of
this came from Kagan Tuncay who followed directly an algorithm
described in a classic book on finite element methods by
Zienkowicz.  Problem is this algorithm no doubt worked fine
in finite element analysis, but the iteration diverged
when and only when the point requested was very close to
and edge of a box skewed in particular form.  Found this only
by looking one by one at failure nodes in a 3d perspective which
led me to the underlying problem in this method.  The new
algorithm has an operation count about 10 times smaller than
the previouvs version and seems to be dead stable.
It uses a much simpler algorithm (ha) in which the transformation
matrix is computed and a change of variables applied to put the
requested point in a standard reference frame.  Then interpolation
by the bilinear shape functions works perfectly even if a point
is located on the edge or event slightly outside a box.