Fix a sign error in the free_surface_transformation method and update documentation to clarify the confusion that caused this.
A confusing sign convention bit me here.  azimuth is clockwise
from north while phi in spherical coordinates is counterclockwise from
the x1 axis.  The free_surface_transformation routine botched this
in a way that caused the radial and transverse outputs to be switched
after the method was called on a 3c seismogram.