Add ifdefs and a special makefile to support api change between 4.9 and 4.10.
I know some think this is a bad idea, but this mechanism allows a transition
for older versions of Antelope.  I needed this for the time being because
I only have 4.9 on the teragrid machine and it is a pain to get them
to do the upgrade so I won't do so until it is necessary to do so.