Fix bugs in GenericGather (receiver array) processing mode.
These changes fix some serious bugs in the previous version that
caused it to not work with GenericGather processing.
Most of the actual changes where to the XcorProcessingEngine object,
but there are changes here to connect to some new procedures
added to fix mistakes in that code.

xsaa is a new extension table that is very similar to xcorarrival.
At some point xcorarrival and this new extension table need to
be unified.  I discovered an error in my design of the xcorarrival
table that for now is handled through xsaa.  Specifically, the
keys in xcorarrival are not right for processing common receiver
gathers because the keys in xcorarrival are not unique in that
mode.