Fix some problems in the load_data routine when driven from a ProcessingQueue.
This fixes a set of bugs uncovered in testing dbxcor with source
array processing.  Small change to a procedure rotate_procedures to
make it more robust in the presence of illegal input.  Previous
version returned a NaN in the emergence angle if the slowness
vector given implied an evanescent wavefield.  Now it returns 0,
which isn't really correct but at least safe.   Suppose it should
be pi/2, but I think that creates too many other problems downstream
in the normal use of this procedure (ray coordinate rotations).