Add new parameter file examples for source array processing.
Setting up dbxcor to run in source array mode requires a drastically different
parameter file set than for the normal mode of teleseismic body wave processing.
To give me or anyone else a hope of figuring this out this supplies a README
and two examples for P and S wave data.  See README_parameterfiles if you
want to know more.

Ultimately, this needs to be spelled out in a document that gives more
details, but the usual time constraints make this the choice for today.
Better than nothing as they say.