Fix an off by one error in generic median function.
The generic median function (template) here had an off vby one error that caused
it to return an incorrect value for even values of the number of elements in
a vector.  One of those infamous C versus fortran indexing problems.  The worst
case was n=2 where this sometimes returned absurd numbers.

This is a low level function used in dbxcor's stacker, so this is a fairly
important fix.  Largely unimportant for USArray data because n=2 never
happens with dbxcor processing and the number of degrees of freedom makes
the off by one error here unimportant.