Add capability of clearing queue of records marked as busy.
The previous version of this program didn't really work as expected.
Had to work through the entire queue clearing records marked as busy.
dbxcor, in particular, does this at present if it crashes during processing
of common receiver gathers (source array processing).  Not important for
teleseismic processing in the standard mode.