Make save method more robust.
For right or wrong this makes the save method for dbxcor more robust.
Now uses a dbaddnull followed by dbputv to stop crashing on duplicate
xcorarrival entries.  In the long run we should revisit the schema
definition for this table, but this will, for example, stop dbxcor
from crashing if you reprocess the same data without clearing
the xcorarrival table.