Add new option to control absolute depth in relocations.
Previously this program had a fundamental instability if the earth
model was drastically wrong.  Depths would often get systematically
skewed up or down if the model had drastically wrong velocities.
this is handled now through a freeze mode that allows one to
automatically select one event in a cluster as a pseudo-calibration
event.