Move comm and deployment extensions to contrib.
Added Range to deployment fields.

Modified Makefile to make it more readable and added in comm
and deployment.

The comm table is still in its "original" format but will soon
be modified to include additional fields and the field sizes will
be changed (expanded).  As no one but the ANF is likely using this
extension, hopefully the soon-to-be-made changes will go by unnoticed.