add qctests and newcomm; modified dlsite and comm extension schemas
qctests is the first pass at a table to track results from various
matlab tests that check a few QC results from SOH data.

newcomm is a temporary extension table that will eventually replace the
current comm table.  newcomm has the same information as comm, but the
field sizes have been extended and additional fields are tracked.  As an
interim step there are also some description changes for dlsite and comm.

Eventually, the record sizes will be changed for dlsite and comm, but as
the ANF is the only user of these schema extensions (as far as I know),
there should be little affect for down stream users.  However, I have
been known to underestimate difficulties in the past...