New data object to handle event catalogs in a way that allows fast indexing and comparison with existing events.
Sometimes it is useful to have an indexed source geometry array similar
to the concept of as seismic array indexed by station names.  Indexing of
events, however, is a drastically different thing.  This object defines
an interface to access this abstract concept of a source array.  The
actual implementation uses an STL map object with a novel indexing
scheme.  The map key is the Hypocenter itself and comparisons are
made through a space-time distance calculation.  This effectively
builds an associative array of Hypocenters in time order.  Equality is
defined in the concept used in the STL map object called weak order.

This object also has a interesting abstraction of the concept of
developing a subset of events.  This is done through a template
and an unspecified "Predicate" function object used in STL
generic algorithms.  The basic idea is the Predicate can do
anything as long it can answer yes or not when given a particular
Hypocenter object.  At the moment no specific subsetting Predicate
has been implemented, but the interface has this defined.

This is an initial checkin of this object.  Most, but not all methods
in the object have been fully tested.