New version to enhance support for source array processing.
This version implements a new recursive feature for source array processing.
Found from experience that source array gathers often have multiple potential
suites of waveforms that can be correlated successfully.  To utilize this when
the program is run in GenericGather mode the analysis becomes recursive.  That is,
when one group of data is correlated and saved, the display is refreshed to
show only data that was not saved as part of the previous analysis, although still
from the same parent gather.

At the same time I generalized the method used to generate a plot title to allow
one to display something useful when running in GenericGather mode.  Previous
version just posted a generic (literally) string.  Now the user can specify
what ensemble attributes are to be posted to each plot title.  There is a test in
the session_manager constructor to make sure the attributes requested for the
title are required to be loaded from the database.