Add manual picking method.
This is hopefully the end of feature creep in this program.  This version adds
a new button that enables manual picking on the main display window.  This is
always on, but is of most use in source gather processing.  Found many instances
where it would be nice to fix an analyst's pick in the context of other waveforms
from the same volume of the earth.  This adds that capability.