Recover important datalogger event message from .*/log packets and store them in a database (dlevent) table.
This code uses similar logic (and the same pf file) as the
logs2dlevent program.  The messages are recovered via some rather
ugly grep commands that look for matching phrases that are descibed
in the pf file.  As caveated in the man page, this has only been
tested on logs output from q3302orb and any grep of logs from another
program is prone to problems if the freeform log output changes.