Add new capabilities to MultichannelCorrelator to improve performance of dbxcor.
These changes add two new options for running the MultichannelCorrelator.
(1) Correlate only, meaning data are shifted but stack is set to
the data trace passed in originally as a reference.  This has been
found to be useful in source array processing to find unambiguous
cluster events in a large ensemble.  (2) stack only.  This is pretty
much the opposite of correlate only.  That is, cross-correlation functions
are computed but data are not time shifted.  The stack is then
produced from data without changing any time alignments.  This is
useful in a variety of contexts, but the primary reason was a need
to easily use data that stubbornly refuse to not undergo a cycle skip.
Another is that this is a very simple and clean way to construct
a suite of extended beams over arbitrary time gates with dbxcor.
First pass would effectively set the arrival times and the second pass
could made in this mode to construct an interactively chosen wavelet
that was coherent across the entire array.  Note this requires also
setting the option to align on picks, not on theoretical arrival times.