Extensive reworking of the GUI for this program.
This change is a complete overhaul of the gui or this program.
(1) Add a new option item for MultichannelCorrelator mode.  See updates
to man page on how to use this feature.
(2) Many former bottom row buttons were moved to an edit menu.  This was
done for three  reasons.  First, it provides a more logical arrangment of
functions.  Second, it avoids a problem with some motif implementations
(i.e Sun) that did not handle geometry management correctly with the long
string of buttons.   This probably could have been corrected other ways, but
this solved this problem as well.  Finally, and perhaps most importantly,
it allowed defining keyboard accelerators for the most heavily used methods.
As I read the Motif manual these can only be used for top menu items.
Early experience with this new gui indicates and dramatic improvement
in efficiency with one hand using keyboard accelerators and the other hand
doing the mousing.
(3) Man page is updated to reflect the changes, although I don't guarantee
it is doesn't contain some relics I missed.