Fix ambiguous handling of zero slowness vectors.
There is an ambiguity in zero length slowness vectors wrt azimuth.
The azimuth and baz methods of this simple object use atan2 which
yields an NaN in this situation.  Fixed in a general way by adding
a private variable set with a new parameterized constructor that
sets the default angle for this vertical incidence special case.
Added operator = because it seems private variables aren't copied
by the compiler generated operator=.

ray1d is a minor fix of a dumb error found in passing with operator=