Track TA stations which have been transitioned to EARN status or "adopted" by other regional networks.
This is the first version of this schmea extension.  It is subject
to change!

Look for a "mk_dbops" script to be checked in sometime in the next week.
This script will update/modify the comms, deployment, adoption, and
dlsite tables which are used at the ANF for the TA project.  Hopefully
it will save the rt system operators who use these tables from
countless tedious updates via dbe to these tables.