Suite of schema extension changes, conversion utilities, and a program to update these tables as used by the ANF.
NOTE:  If you have used the comm, deployment, or dlsite extension
tables prior to January 2009, you will have to convert your pre-existing
tables to this update format.  Conversion scripts (convert_comm,
convert_deployment, convert_dlsite) are provided to help with this task.
Any in-house software that retrieves information from these tables should
be checked.  The key differences:

	- deployment table now uses vnet rather than net.  The field
	  size is expanded.
	- comm and dlsite table now have extended field size for provider.
	- comm table has additional fields for dutycycle, power, equip_type.

Also included in this check-in is some additional documentation about the
schema extensions.

Finally, to my great relief as a maintainer of 750+ station that have
contributed to the TA, the mk_dbops script will simplify the update of
the adoption, comm, dlsite, and deployment tables.  Previously, all
records had been added/modified by dbe or vi which led to errors of
with omission of information, typos, or mismatch between the tables.

This is a work in progress, and additional updates will likely be
occurring in the upcoming weeks (i.e. station based subsetting using
not yet checked in features of q330_location).