Review mass positions based on channel lead defined in calibration.
This takes care of the 6 channel q330 case where a Trillium is
installed on "Sensor A" and no sensor on "Sensor B".  Prior to this
update, when Sensor B returned its default values of 2.0V, and a
Trillium with a threshold of 2.0V was installed on Sensor A, the
sensor would be determined as "over limit" and sent erroneous MRCs.

The script will now review the calibration table for a lead value and if
there is a lead value, and the channel subset finds a lead value matching
either in the sensor_lead of the pf file, only that sensor will have it's
mass positions reviewed.

The man page has been updated and additional parameters were added to the
parameter file.