Moved the datalogger and sensor parameter files from the BRTT source tree to contrib.
I also added the new originator and last_modified
parameters to the files (from Niko) already in this directory.

And I added a README suggesting a little bit how to add files here,
and suggesting corrections should go to the author.

This README probably needs to have more added, and perhaps
there needs to be some document describing the whole process in more detail.