Move the variants of antelopemake which are intended to allow using some combination of GNU tools to compile Antelope programs under solaris into contrib.  Since BRTT doesn't use these at all, we can't maintain them.
These fragments are dependent on some fragments which are retained in
the Antelope core, and which in the future will be installed in $ANTELOPE/data/makes.
antelopemake-gcc-mix and antelopemake-gcc are created by combining and
editing the files checked in here, and two files from $ANTELOPE/data/makes.

If you need the latter, please contact danq@brtt.com