New program.  Alternative to trexcerpt.  Does some things trexcerpt can't do and doesn't do some things trexcerpt can. Main purpose is to produce a regular gather with a common sample rate using wfprocess to index the waveforms so we avoid the infamous overlapping waveforms problem. It also can do three-component bundling which can provide input data to speed processing with dbxcor.
This has been in my development tree for months, but it is time to put
it into the distribution.  The main driver is I am going to use it in
the USArray data processing workshop in a couple weeks.   Read the man
page if you see this and are interested in a trexcerpt alternative.