Implement automatic byte swapping on output.
Previous version created a lot of problems with mixed endian data.
Because a gclfield is stored in two pieces (grid is considered
independent of the contents) one could previously get mixed endian
data in the two pieces.  Because the gclfield table does not store
a datatype attribute, this created havoc.  I chose to implement byte
swapping on output rather than modify the schema because I have way
too many existing tables that would have to be repaired.

Also add a new procedure to decimate a GCLgrid.  This previously was
in two different components of my plane wave migration code, but it
belongs here.