Fix to make work on macos.
This may be a gcc version skew issue, but this compiled on linux but failed on macos.
Linked to a temporary variable used in a function call.  Some oddity linked to use of
the C++ reference caused this to not compile.  This is a workaround with a tiny
memory overhead.