Regularize error handlers.
Previous version didn't implement inheritance properly for the suite
of error objects I (incorrectly) created as this library evolved.  The
problem is a subtle issue about references to base classes in C++ and how
it impacts how one should do error handlers.  Previous version was a mishmash.
This version now has this simple scheme.  All users of this library should
surround calls to anything inside it with a catch phrase with a reference
(not the object a reference) to a SeisppError object.  All routines post
to an internal message buffer that can be retrieved with the what method
(like std::exception, but SeisppError does not inherit std::exception
 as it proved to have some odd issues with it's destructor)  or the
log_error() method.  The plethora of error objects still exist, but they
can all now be handled by a single catch for a SeisppError reference.

Although the changes made here are pretty small they effect almost the
entire library.  I went through the entire library and cleaned up my
incorrect uses of copies versus references as appropriate.