Bunch of changes to improve performance of dbxcor and to fix a couple bugs.
Main changes are to XcorProcessingEngine with some cascading issues
that required additions/fixes in library routines.  cvs is lying as
only about five files were modified.

Also adding a couple of new procedural function files unrelated to
dbxcor.  SimpleWavelets has some trivial functions to produce
a Ricker wavelet and or a Gaussian wavelet.  Sparse convolve is
a useful procedure for sythetics made up of a small number of
impulses, which is what I made it for.