Collect external bulletins for distribution as /pf/orb2dbt packets.
This is the initial check-in of my long promised script that will
collect and parse bulletins from a wide variety of regional
networks and institutions and make them available for Antelope users
in an easy to use format.  This should alleviate some pains that
regional network operators have had over the years in maintaining
their own collectors and parsers for various regional network
and USGS/NEIC bulletins.  The suggested workflow is that this
script will be run at the ANF as a service to the community and
remote users can orb2orb the /pf/*/orb2dbt packets of interest
to their local institutions.  The desired /pf/*/orb2dbt packets
can then be read to local disk with orb2dbt (or multiple orb2dbt
instances).

I checked in a companion to this script, a new "bulletins" rtdemo,
to the BRTT source code earlier today.  The new rtdemo shows
how the pf packets can be collected from anfexport.ucsd.edu:bulletin
and written to local disk.

Please contact me if you find that your favorite bulletin is not
currently supported and I will see what can be done to add it in.