New error flag on returned JSON object
The system will return some text to the browser in case there is a
problem. The setData and the printData functions on the client are
updated to look for this element and trigger alerts to the user. The
setData function will not crash if there is no 'data' element on the
returned object.