Fix bug on eventdata.py for function phases(). Update null_vals().
Now the query for phases will test the existence of table assoc and will
try to join that with table arrival. If no assoc table is detected the
system will append an 'i_' label to the phase string.
Minor update on the function to query for null values. The system will
only query the database once and then the values will be cached on the
variable self.nulls on the respective CLASS.