ANF specific QEP (Quanterra Environmental Processor), setra and NCPA dbbuild batch files.
Each QEP infrasound channel (1, 2, or 3) has a unique sensitivity value for
the attached sensors.  In the TA setup, one QEP channel has a setra278
attached, the 2nd QEP channel has an NCPA sensor, and the 3rd QEP channel
may have a 2nd NCPQ sensor.  The meteorological data is collected with an
_EP network code and requires the qep_soh_only dbbuild pf file.