php: update orb interface to new 5.2 orb data structures
The underlying orbclient and orbstat structures changed in their
internal structure. The removal of the 'priority' entry in the orbclient
structure in particular was breaking compilation.

Additional functions were created to retrieve values for the new entries
in orbclient and orbstat. These were determined by diffing
orb_protocol.h between 5.1 and 5.2