Fix broken and soon to be deprecated xpls suffix
Antelope 5.4 breaks the xpls Stand-alone Perl script file type by not
having the licensing capabilities available at run time unless the full
Antelope environment is sourced. This is problematic because xpls files
are supposed to be able to run on their own, for example as nagios
checks or CGI-BIN scripts in a web server.

Since BRTT has decided that they are no longer going to support the xpls
extension, it has been moved to contrib.

The fix for the licensing issue is to "require" one of the Antelope perl
modules rather than "use". This causes the library to load at script
compile time, and thus brings in the licensing.

Note that the method used to force the licensing to load - requiring the
'sysinfo' module - does not seem to have any ill-effect if a use sysinfo
statement is found later in the script, so existing scripts shouldn't be
affected.