New variables PMAN1 and PMAN3
Instead of overriding the MAN1 and MAN3 variables and associated rules
from $ANTELOPEMAKE, use new variables PMAN1 and PMAN3 that are designed
to take .xpls and .pm files as their values. The pod2man(1) command is
run on the input files to generate man pages with intermediate .pd.1 and
pd.3p extensions (so that you can preview the files before they are
installed), but then are installed with the proper .1 or .3 extensions
in the proper man directory.